requirements.txt
django==1.11
djangorestframework==3.6.3
gunicorn==19.6.0
psycopg2==2.6.2
requests==2.11.1
django-storages==1.6.3
boto==2.48.0
django-markdown2==0.3.0
markdown2
docker
python-dateutil

django-filter
django-anymail
pytz

raven
newrelic

coverage
faker
ipdb
ipython
